src/fetch_weather.py
"""
Weather data fetcher module
Fetches weather data from various APIs
"""

import os
import json
import requests
from datetime import datetime
from dotenv import load_dotenv

load_dotenv()

def main():
    """
    Main function to fetch weather data
    """
    print("Fetching weather data...")
    
    # TODO: Add your weather API implementation here
    # This is a placeholder that creates sample data
    
    weather_data = {
        "timestamp": datetime.now().isoformat(),
        "location": "Florida",
        "temperature": 75,
        "conditions": "Partly Cloudy",
        "humidity": 65,
        "wind_speed": 10,
        "forecast": "Sunny with occasional clouds"
    }
    
    # Save to data directory
    os.makedirs("data", exist_ok=True)
    with open("data/weather_data.json", "w") as f:
        json.dump(weather_data, f, indent=2)
    
    print("Weather data saved to data/weather_data.json")
    return weather_data

if __name__ == "__main__":
    main()