requirements.txt
requests
openai
python-dotenv
jinja2
markupsafe
pytz